project status, new project, add project subcommands
