--- src/commands/projects/SetProjectStatus.ts
@@ -1,44 +1,0 @@
-import { SlashCommandBuilder, PermissionFlagsBits, ForumChannel, CommandInteractionOptionResolver, ChatInputCommandInteraction } from "discord.js";
-import CrossingGuardBot from "../../CrossingGuardBot";
-import { ProjectStatus } from "../../ProjectStatus";
-
-const data = new SlashCommandBuilder()
-    .setName("setprojectstatus")
-    .setDescription("Set a project's status")
-    .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
-    .addStringOption(option =>
-        option.setName("project_name")
-            .setDescription("The name of the project")
-            .setRequired(true))
-    .addStringOption(option =>
-        option.setName("status")
-            .setDescription("The current status of the project")
-            .setRequired(true)
-            .addChoices(
-                { name: "Playable", value: "0" },
-                { name: "In Development", value: "1" },
-                { name: "Archived", value: "2" },
-                { name: "Hidden", value: "4" }
-            ));
-
-async function execute(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("project_name");
-    const status = interaction.options.getString("status");
-
-    if (!projectName || !status)
-        return;
-
-    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
-        if (!project) {
-            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
-            return;
-        }
-
-        project.status = +status;
-        CrossingGuardBot.getInstance().database.saveProject(project);
-        interaction.reply({ content: `${project.displayName}'s status set to ${ProjectStatus[+status]}`, ephemeral: true });
-    });
-
-}
-
-export { data, execute };
